pyproject.toml
[project]
name = "granitedoclingtesting"
version = "0.1.0"
description = "testing the granite docling model with pdfs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling",
    "dotenv>=0.9.9",
    "google-genai>=1.45.0",
    "onnxruntime-gpu>=1.23.0",
    "openai>=2.4.0",
    "torch",
    "torchvision",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
